Basic functionality and demonstration
The major components are heaps and cells.  Ancillary utility components,
cell holder, root and reference are provided for convenience.

Heaps may be used as template parameters or via a type erased
(exact::TypeErasedHeap) implementation which invokes all operations via
a virtual table (exact::Heap).  A poorly optimized but functional heap
implementation is provided (exact::NonMovingMarkSweep).

A cell is a class template (exact::Cell<T>) that is a relocatable buffer
of adequate size to fit an object of the template parameter.

For convenience, exact::Root, exact::RootUtil, exact::Reference, and
exact::ReferenceUtil are provided to automatically manage roots and
reference, and to create them.